feat: regras de autorização por dono e tipo de usuário

+++ core/views.py
@@ -2,12 +2,17 @@ from django.shortcuts import render, get_object_or_404, redirect
 from .models import Demanda
 from .forms import DemandaForm
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
 
 def home(request):
     demandas = Demanda.objects.filter(status="ABERTA").order_by("-criada_em")
     return render(request, "core/index.html", {"demandas": demandas})
 
+@login_required
 def criar_demanda(request):
+    if request.user.tipo != "CLIENTE":
+        raise PermissionDenied
     if request.method == "POST":
         form = DemandaForm(request.POST)
         if form.is_valid():
@@ -24,9 +29,12 @@ def detalhe_demanda(request, pk): # pk é o id da demanda
     demanda = get_object_or_404(Demanda, pk=pk)
     return render(request, "core/detalhe.html", {"demanda": demanda})
 
+@login_required
 def editar_demanda(request, pk):
     demanda = get_object_or_404(Demanda, pk=pk)
-
+    if demanda.cliente != request.user:
+        raise PermissionDenied
+    
     if request.method == "POST":
         form = DemandaForm(request.POST, instance=demanda)
         if form.is_valid():
@@ -38,9 +46,11 @@ def editar_demanda(request, pk):
 
     return render(request, "core/criar_demanda.html", {"form": form})
 
+@login_required
 def excluir_demanda(request, pk):
     demanda = get_object_or_404(Demanda, pk=pk)
-
+    if demanda.cliente != request.user:
+        raise PermissionDenied
     if request.method == "POST":
         demanda.delete()
         messages.success(request, "Demanda excluída com sucesso!")
